+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+<!-- eslint-disable markdown/no-missing-label-refs -->
+
 # <img src="https://matterbridge.io/assets/matterbridge.svg" alt="Matterbridge Logo" width="64px" height="64px">&nbsp;&nbsp;&nbsp;Matterbridge hass plugin changelog
 
 All notable changes to this project will be documented in this file.
@@ -9,6 +11,29 @@ If you like this project and find it useful, please consider giving it a **star*
 ## Possible issue upgrading the plugin
 
 > WARNING: The domains button, remote and media_player include an OnOff cluster. This will not make it possible to merge the entities on the same endpoint: if you have Alexa or Google you may want to either black list those domains or to split their entities.
+
+## [1.2.1] - 2026-05-01
+
+### Added
+
+- [light]: Add discovery of dimmer lights when they are `unavailable` in Home Assistant. Thanks DrFate09.
+- [cache]: Add `StateCache` class to store the device states when they turn to unavailable in Home Assistant. This allows to create correct lights even if they are unavailable on Home Assistant.
+- [eslint]: Enforce @typescript-eslint promise rules.
+
+### Changed
+
+- [package]: Update dependencies.
+- [package]: Bump package to `automator` v.3.1.7.
+- [package]: Bump `eslint` to v.10.3.0.
+- [package]: Bump `typescript-eslint` to v.8.59.1.
+- [eslint]: Remove `eslint-plugin-promise` (not actively maintained) and add optional @typescript-eslint promise rules.
+- [package]: Remove `overrides` that was necessary for eslint-plugin-promise.
+- [package]: Remove unused Vitest setup.
+- [eslint]: Add `eslint` v.2.0.0 config.
+- [prettier]: Add `prettier` v.2.0.0 config.
+- [jest]: Add `jest` v.2.0.0 config.
+
+<a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
 
 ## [1.2.0] - 2026-04-24